refactor: move withdraw logic to one place
